moved methods around; renamed ecosystem.py to environment.py
